+++ platform/marvell-arm64/sonic-platform-asterfusion/x202p-16s/asterProcess/main.py
@@ -34,21 +34,21 @@ class GlobalThread(common.threading.Thread):
             self.onMessage(message)
 
     def onMessage(self, message):
-		"""
-		Commands:
-			uninstall	: Uninstall platform drivers
-		"""
-		if len(message.command) < 1:
-			result = self.onMessage.__doc__
-		else:
-			if message.command[0] == 'uninstall':
-				common.RUN = False
-				doUninstall()
-				result = 'Success'
-			else:
-				result = self.onMessage.__doc__
-		if (message.callback is not None):
-			message.callback(result)
+        """
+        Commands:
+            uninstall	: Uninstall platform drivers
+        """
+        if len(message.command) < 1:
+            result = self.onMessage.__doc__
+        else:
+            if message.command[0] == 'uninstall':
+                common.RUN = False
+                doUninstall()
+                result = 'Success'
+            else:
+                result = self.onMessage.__doc__
+        if (message.callback is not None):
+            message.callback(result)
 
 class messageObject(object):
     def __init__(self, command, callback):
@@ -141,7 +141,8 @@ def doInstall():
     for o in i2c_topology_dict:
         status, output = common.doBash("echo " + o['driver'] + " " + o['address'] + " > " + common.I2C_PREFIX + o['bus'] + "/new_device")
     status, output = common.doBash("sysctl -w net.mpls.platform_labels=1048575")
-    status, output = common.doBash("echo -2 | tee /sys/bus/i2c/drivers/pca954x/*-00*/idle_state")
+    if os.path.exists("/sys/bus/i2c/drivers/pca954x"):
+        status, output = common.doBash("echo -2 | tee /sys/bus/i2c/drivers/pca954x/*-00*/idle_state")
     print("end do Install ...")
     return
