+++ platform/marvell-arm64/sonic-platform-asterfusion/x206p-48s-ptp/asterProcess/main.py
@@ -148,7 +148,8 @@ def doInstall():
     for o in i2c_topology_dict:
         status, output = common.doBash("echo " + o['driver'] + " " + o['address'] + " > " + common.I2C_PREFIX + o['bus'] + "/new_device")
     status, output = common.doBash("sysctl -w net.mpls.platform_labels=1048575")
-    status, output = common.doBash("echo -2 | tee /sys/bus/i2c/drivers/pca954x/*-00*/idle_state")
+    if os.path.exists("/sys/bus/i2c/drivers/pca954x"):
+        status, output = common.doBash("echo -2 | tee /sys/bus/i2c/drivers/pca954x/*-00*/idle_state")
     print("end do Install ...")
     return
